TST: unified example files
